Aleksey Baliuk: refactor name
