+++ src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { AuthGuard } from './commonGuards/auth-guard';
+import { AuthGuard } from './common-guards/auth-guard';
 
 const routes: Routes = [
   { path: 'auth', loadChildren: './modules/auth/auth.module#AuthModule' },
