+++ src/app/common-guards/auth-guard.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
 import { Observable } from 'rxjs';
-import {GlobalAuthService} from '../commonServices/global-auth.service';
+import {GlobalAuthService} from '../common-services/global-auth.service';
 
 @Injectable({
   providedIn: 'root'
